requirements.txt
flask
python-dotenv
groq
werkzeug
edge-tts